[PAPI] Implement generic MPI interface

+++ src/papi/papi_wrapper.sh
@@ -9,8 +9,10 @@ PAPIDIR="$SRCDIR/papi"
 . "$PAPIDIR/../utils/general_utils.sh"
 . "$PAPIDIR/../utils/print_utils.sh"
 . "$PAPIDIR/../utils/array_utils.sh"
+. "$PAPIDIR/../utils/mpi_utils.sh"
 
-PAPI_PROFILER="$PAPIDIR/papi_profiler.o"
+PAPI_PROFILER_NAME="papi_profiler.o"
+PAPI_PROFILER="$PAPIDIR/$PAPI_PROFILER_NAME"
 
 
 # Returns 0 if papi is available, 1 otherwise.
@@ -27,28 +29,43 @@ papi_available() {
     return 0
 }
 
+# Compile profiler binary if needed.
 _compile_papi_profiler() {
-    # Check if the binary exists.
-    if [ -f "$PAPI_PROFILER" ]; then
-        # If exists, check for executable rights and rebuild if not.
-        if [ -x "$PAPI_PROFILER" ]; then
-            return
-        else
+    # Return if profiler exists and is executable.
+    if [ -x "$PAPI_PROFILER" ]; then
+        verbose_echo print_info "$PAPI_PROFILER_NAME is already compiled."
+        return
+    fi
+
+    # Recompile profiler with rank 0, and let other ranks wait.
+    if [ "$RANK" -eq "0" ]; then
+        verbose_echo print_info "Compiling papi_profiler.c.."
+
+        # Remove profiler binary if it exists but isn't executable.
+        if [ -f "$PAPI_PROFILER" ]; then
             rm "$PAPI_PROFILER"
         fi
-    fi
 
+        # Compile the code.
+        if ! cc "$PAPIDIR/papi_profiler.c" "$PAPIDIR/papi_component.c" \
+                "$PAPIDIR/papi_event.c" -o "$PAPI_PROFILER" -lpapi -Wall; then
+            print_error \
+                "Error while compiling $(basename "$PAPI_PROFILER"), exiting.."
+            exit 1
+        fi
 
-    # Compile the code.
-    if ! cc "$PAPIDIR/papi_profiler.c" "$PAPIDIR/papi_component.c" "$PAPIDIR/papi_event.c" -o "$PAPI_PROFILER" -lpapi -Wall; then
-        print_error "Error while compiling $(basename "$PAPI_PROFILER"), exiting.."
-        exit 1
-    fi
-
-    # Make binary executable.
-    if ! chmod +x "$PAPI_PROFILER"; then
-        print_error "Error during 'chmod +x $(basename "$PAPI_PROFILER")', exiting.."
-        exit 1
+        # Make binary executable.
+        if ! chmod +x "$PAPI_PROFILER"; then
+            print_error \
+                "Error during 'chmod +x $(basename "$PAPI_PROFILER")', exiting.."
+            exit 1
+        fi
+    else
+        # Rank >0 wait till profiler exists.
+        verbose_echo print_info "Waiting for rank 0 to compile profiler.."
+        while [ ! -x "$PAPI_PROFILER" ]; do
+            sleep 0.2
+        done
     fi
 }
 
@@ -327,14 +344,29 @@ papi_profile() {
         return
     fi
 
-    verbose_echo print_info "Events:\n$events"
-    verbose_echo print_info "Units:\n$units"
+    verbose_echo print_info "Events:
+$events"
+    verbose_echo print_info "Units:
+$units"
 
     # Make sure the papi_profiler is updated and compiled.
-    verbose_echo print_info "Compiling papi_profiler.c.."
     _compile_papi_profiler
 
-    # Profile binary with supported events.
-    verbose_echo print_into "Executing profiler"
-    "$PAPI_PROFILER" "$events" "$units" "$@"
+    # Compute the number of nodes using all ranks.
+    num_nodes=$(mpi_get_num_nodes)
+
+    # PAPI uses RAPL which profiles the entire node, thus only the first rank
+    # per node needs to profile.
+    if [ "$LOCAL_RANK" -eq "0" ]; then
+        # Collect measurements with PAPI.
+        verbose_echo print_info "Executing profiler"
+        output=$("$PAPI_PROFILER" "$events" "$units" "$@")
+
+        # Split output in events and energies, then merge MPI ranks.
+        events=$(printf '%s\n' "$output" | awk '{print $1}')
+        energies=$(printf '%s\n' "$output" | awk '{print $2}')
+        mpi_gather "$MPI_MODE" "$num_nodes" "$events" "$energies"
+    else
+        verbose_echo print_info "Local rank > 0, so exiting.."
+    fi
 }
